+++ src/lib.rs
@@ -1,10 +1,5 @@
 extern crate mjolnir_api;
 
-// mod agent;
-mod plugins;
-
-pub use plugins::PluginEntry;
-
 #[cfg(test)]
 mod tests {
     #[test]
